python_api/requirements.txt
flask
langchain
langchain-community
langchain-anthropic
anthropic
flask-cors
python-dotenv